+++ src/components/token/index.js
@@ -1,0 +1,5 @@
+export {
+  EuiToken,
+  TYPES as TOKEN_TYPES,
+  SIZES as TOKEN_SIZES
+} from './token';